NB: Improve Ajax server>clientS push reliability mechanism
Prior design had a fault that exposed a ~300ms window during which
pushed events could be dropped on the floor.

When hk-ch puller pulled an active-but-soon-to-close-hk-ch, it'd
mark the client's uuid as satisfied despite its send actually being
a noop.

The solution here is to move the `send!` itself into the reliability
mechanism so that:
1. Http-kit's `send!` result can be inspected before acknowledging a
   client uuid as satisfied.
2. Any pulled hk-ch's are consumed immediately, dramaticly decreasing
   the probability of a channel closing before consumption.