Make channel socket available in event-msg-handler
To allow a message handler to be used without reference to a singleton
channel socket instance, pass channel-socket information to the
event-msg-handler function.

On the server side, the :channel-socket key is added to the event map
passed as the first argument, and the ch-recv argument is removed (as it
is now available under the :channel-socket.

On the client side, the event-msg-handler signature is changed to match
the server side more closely, taking a single map argument that has
:event and :channel-socket keys.

The new signature is used by the new start-message-router! functions.

The old signature is still used by the start-chsk-router-loop!
functions.