[#105] EXPERIMENTAL: enable clients to distinguish between automatic + manual reconnects
Clients receive `[:chsk/state <new-state>]` `event-msg`s on their `:ch-recv`
core.async channel when the status of their channel socket changes (e.g. the
socket opens or closes).

This commit adds a `{:requested-reconnect? true}` entry to the relevant
`<new-state>` values. It wasn't previoualy possible to watch for chsk state
changes and know on a state change event if the change was due to an
automatic reconnect, or due to an explicit `chsk-reconnect!` request.

IMPLEMENTATION NOTE:
`chsk-reconnect!` was previously NOT triggering an `{:open? false}`
client-side state change for Ajax connections. This wasn't intentional,
so I took the opportunity to clean up the Ajax handshake procedure.

Before:
Ajax GETs would trigger a server-side handshake response iff the
request's client-id was "new". The client had no ability to _request_
a handshake response.

Now:
Ajax GETs trigger a server-side handshake if the request's client-id
is "new". The client also has the ability to _request_ a handshake
response whenever it likes. It chooses to do so whenever it finds
itself in a `{:open? false}` state.

The new behaviour should also be more robust to Ajax disconnections
generally.