Refactor client/server assumptions (further prep for #168)
Further decoupling assumption that clj<=>server and cljs<=>client

As of this commit:
  server e/o #{clj cljs}
  client e/o #{cljs}

Further work would be necessary to allow a clj client but shouldn't
be too hard (PRs welcome to get this started!).

Big thanks to Matthew Molly (@whamtet) and Andrew Phillips (@theasp)
for getting the ball rolling on this!