Make the CSRF token location configurable
Rather than having the CSRF location hardcoded, allow specifying
`:csrf-path` for `make-channel-socket-server!`, which is passed to
clients during handshake, describing where a client is to place the CSRF
token for posts, to match the configuration of the server.  This allows
the server to use the standard middleware without any options for both
ring-anti-forgery and csurf when you use a value of `[:headers
:X-CSRF-Token]`.  The default is `[:params :csrf-token]` to maintain
compatibility with previous versions.

The example server has been upgraded to use `[:headers :X-CSRF-Token]`,
and the example client checks to see if there was an error to determine
login success or not.