[#230] GC non-terminating websocket conns
If a WebSocket connection is closed without normal termination (e.g. as caused
by sudden loss of power, airplane mode, etc.) - it can hang around in conns_
for an extended period of time until the underlying TCP connection is identified
as dead.

This simple mod allows Sente's server to perform WebSocket connection GC for
clients that haven't communicated with the server w/in a defined amount of time
(must be > client :ws-kalive-ms).

Big thanks to @altV for helping to catch + diagnose this issue.