Nb server interface housekeeping, server now controls `websocket?` flag
In particular, #236 identified that Aleph doesn't provide the
necessary `{:websocket? true}` flag. This may actually have been
the cause of trouble with #225, still need to confirm.

As of this commit, servers now determine entirely by themselves
whether or not to treat a particular request as a WebSocket upgrade.
This should help prevent split-brain issues caused by Sente+server
disagreement and make it easier to identify buggy server
implementations, etc.