[new] Allow WebSocket constructors to delay connection
Previously:
  WebSocket constructors returned a connected WebSocket.

After this commit:
  WebSocket constructors may return:
  - nil     (=> no WebSocket support), or
  - a delay (=> deref to return a connected WebSocket)

In other words, we've decoupled construction and connection.

The major advantage of this change is that it's now possible
for the default `ChAutoSocket` (:auto type chsk) to distinguish
between two kinds of errors:

  1. The platform doesn't support WebSockets (=> fall back to Ajax)
  2. The platform does    support WebSockets, but current
     connection attempts are failing due to transient reasons like
     the internet being temporarily unavailable (=> keep retrying)

Note: the new capability is implemented in such a way that we won't
break pre-existing custom constructors. While constructors should
ideally now return an unconnected delay, they MAY still directly
return an already-connected socket.