[nop] [#437] Add info re: choice to wrap client-side pings
Client-side `[:chsk/ws-ping]` events are wrapped to
`[:chsk/recv [:chsk/ws-ping]]` when `wrap-recv-evs?` is true.

In fact, it'd be better if they were NOT wrapped - even when
`wrap-recv-evs?` is true.

While `:chsk/ping` is indeed pushed from server, `:chsk/recv` is
in practice reserved for user-level events.

However I'm NOT changing this because:

  - It'd be a breaking change for users with `wrap-recv-evs?` enabled.
  - The current recommendation (and default) is anyway to disable
    `wrap-recv-evs?`, so the benefit in changing the behaviour now would
    be small.

This commit just documents the oddity for reference.