Pass CSRF token using Sec-WebSocket-Protocol header
The proposed change tries to respect other values in the
Sec-WebSocket-Protocol header, that the caller may have set in the
headers option when calling make-channel-socket-client!

Use the pre-ws-handshake option introduced in http-kit commit to add
the Sec-WebSocket-Protocol response header in the websocket
upgrade. Some browsers (like Google Chrome and Google Chrome derived
ones) fail to complete the upgrade if the upgrade response doesn't
include the Sec-WebSocket-Protocol with one of the sub-protocols
proposed in the request.

[Re: 465]